TS integration tests are complete

+++ commands.py
@@ -195,7 +195,7 @@ class setup_timeseries(bucket_type_commands, Command):
                     user varchar not null,
                     time timestamp not null,
                     weather varchar not null,
-                    temperature float,
+                    temperature double,
                     PRIMARY KEY(
                         (geohash, user, quantum(time, 15, m)),
                         geohash, user, time
